requirements.txt
invoke==2.2.0
matplotlib==3.10.0
scikit_learn==1.6.0
torch==2.5.1
typer==0.15.1
hydra-core==1.3.2
pytest==8.3.4
torch==2.5.1
typer==0.15.1
omegaconf==2.3.0
